Create a constants subpackage in item, and move type to it
