+++ clock/storage/data_source/data_sources/sqlite/sql/item/constants/__init__.py
@@ -1,0 +1,5 @@
+"""
+The items defined here can have very few values.
+
+All of the values are defined as constants and they should be used instead of instantiating them directly.
+"""
